refactor: implement config class from utils
