+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@secjs/logger",
-  "version": "1.2.7",
+  "version": "1.2.8",
   "description": "",
   "license": "MIT",
   "author": "João Lenon",
@@ -25,7 +25,7 @@
     "@secjs/config": "1.1.3",
     "@secjs/env": "1.2.8",
     "@secjs/exceptions": "1.0.4",
-    "@secjs/utils": "1.5.8",
+    "@secjs/utils": "1.7.6",
     "@types/debug": "4.1.5",
     "@types/jest": "27.0.1",
     "@types/node": "14.14.37",
