+++ src/Drivers/ConsoleDriver.ts
@@ -1,6 +1,6 @@
-import { Config } from '@secjs/config'
-import { Color } from '../utils/Color'
-import { format } from '../utils/format'
+import { Config } from '@secjs/utils'
+import { Color } from '../Utils/Color'
+import { format } from '../Utils/format'
 import { DriverContract } from '../Contracts/DriverContract'
 
 export interface ConsoleDriverOpts {
