+++ src/Utils/global.ts
@@ -1,0 +1,11 @@
+import { Log as LogInstance } from '../Log'
+
+export {}
+
+declare global {
+  const Log: LogInstance
+}
+
+const _global = global as any
+
+_global.Log = LogInstance
